requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
tqdm==4.64.1
fake_useragent==0.1.11
tabulate==0.9.0
schedule>=1.1.0

# Google Sheets API
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Selenium for deletion check
selenium>=4.0.0
webdriver-manager>=4.0.0
